rework the whole config parsing and translation to actions.
The current status is a bit crazy, because it tries to create a generic configuration for composable actions as opposed to the very opinionated approach from before, while still trying to stay compatible to all previous configurations. Therefore there is a bit of weird wrapping of all non-map top-level structures, and the hardcoded order of actions. We'll get rid of that eventually when we are ready for a breaking change in config.